requirements.txt
numpy
networkx
antlr4-python3-runtime